server/requirements.txt
cohere==5.13.4
Flask==3.1.0
Flask_Cors==5.0.0
python-dotenv==1.0.1
gunicorn
google.generativeai